moved sboxoverlay to a new file, to clear navbar. Added sample text under it. TO DO: beautify it
